🚚move: header and footer to app folder (#918)
