--- app/components/layout/header.module.css
@@ -1,19 +1,0 @@
-.tiltShaking:hover {
-  animation: tilt-shaking 0.25s linear infinite;
-}
-
-@keyframes tilt-shaking {
-  0%,
-  100% {
-    transform: rotate(0deg);
-  }
-  25% {
-    transform: rotate(5deg);
-  }
-  50% {
-    transform: rotate(0deg);
-  }
-  75% {
-    transform: rotate(-5deg);
-  }
-}
