+++ app/globals.css
@@ -33,6 +33,10 @@ img {
   animation: stretch 1s ease-out forwards;
 }
 
+.tilt-shaking:hover {
+  animation: tilt-shaking 0.25s linear infinite;
+}
+
 @keyframes arrow-up {
   0%,
   40% {
@@ -93,5 +97,21 @@ img {
   }
   100% {
     width: 100%;
+  }
+}
+
+@keyframes tilt-shaking {
+  0%,
+  100% {
+    transform: rotate(0deg);
+  }
+  25% {
+    transform: rotate(5deg);
+  }
+  50% {
+    transform: rotate(0deg);
+  }
+  75% {
+    transform: rotate(-5deg);
   }
 }
